% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DasyMapR-data-documentation.R
\docType{data}
\name{candidates_addresses}
\alias{candidates_addresses}
\title{Candidates addresses of HOU}
\format{An object of class \code{data.frame} with 794875 rows and 6 columns.}
\usage{
candidates_addresses
}
\description{
Non personal data of HOU candidate students
}
\note{
It is subset of dataset and any use for purposes other than educational is prohbited by the owner (HOU)
}
\author{
HOU
}
\keyword{datasets}

