% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ETRS.R
\docType{methods}
\name{CheckEtrsResolution,numeric-method}
\alias{CheckEtrsResolution,numeric-method}
\title{Checks if the given resolution is INSPIRED compatible}
\usage{
\S4method{CheckEtrsResolution}{numeric}(cell.size)
}
\arguments{
\item{cell.size}{numeric one of the resolutions of the grid that are 0.5m, 1m, 2.5m, 5m, 10m, 25m, 50m, 100m, 250m, 500m, 1Km, 2.5Km, 5Km, 10Km, 25Km, 50Km, 100Km}
}
\value{
TRUE if everything is ok if it is not returns 100km cell-size based grid
}
\description{
The grid is defined as a hierarchical one in metric coordinates to the power of 10.The resolution of the grid is 1m, 10m, 100m, 1000m, 10,000m, 100,000m. Althought EEA recommend grid size of metric resolution in standard size 100 m, 1 km, 10 km and 100 km. Alternatively use 25 m or 250 m for analysis purposes, where standard 100 m or 1 km grid size is not appropriate.
}
\examples{

CheckEtrsResolution(cell.size=1000)
CheckEtrsResolution(cell.size=125)

}

