% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ETRS.R
\docType{methods}
\name{CheckEtrsValidity,Spatial-method}
\alias{CheckEtrsValidity,Spatial-method}
\title{Checks the ETRS validity of a Spatial Object}
\usage{
\S4method{CheckEtrsValidity}{Spatial}(object)
}
\arguments{
\item{ETRS}{: Accepts as an argument a ETRS object}
}
\value{
TRUE if everything is ok
}
\description{
Checks the ETRS validity of a Spatial Object
\code{CheckEtrsValidity()} actually checks if the given object is Spatial
and also if it is projected in ETRS-LAEA
}
\examples{
xc <- round(runif(10, min = 4321000, max = 4323000), 2)
yc <- round(runif(10, min = 3210000, max = 3212000), 2)
xy <- cbind(xc, yc)
# a SpatialPoints object
xy.sp <- SpatialPoints(xy, proj4string = CRS("+init=epsg:2100"))
# CheckEtrsValidity(xy.sp)
xy.sp@proj4string<-CRS("+init=epsg:3035")
CheckEtrsValidity(xy.sp)
}

