% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DasyMapR-data-documentation.R
\docType{data}
\name{CLC2000.ARGOLIDA.RES}
\alias{CLC2000.ARGOLIDA.RES}
\title{Corine Land Cover 2000 seamless vector data}
\format{Spatial Polygons Data Frames}
\source{
\url{http://www.eea.europa.eu/data-and-maps/data/corine-land-cover-2000-clc2000-seamless-vector-database-1#tab-interactive-maps-produced}
}
\usage{
CLC2000.ARGOLIDA.RES
}
\description{
Version 17 (12/2013) - Corine land cover 2000 is the year 2000 update of the first CLC database which was finalised in the early 1990s as part of the European Commission programme to COoRdinate INformation on the Environment (Corine)
}
\note{
EEA standard re-use policy: unless otherwise indicated, re-use of content on the EEA website for commercial or non-commercial purposes is permitted free of charge, provided that the source is acknowledged (http://www.eea.europa.eu/legal/copyright). Copyright holder: European Environment Agency (EEA).
}
\author{
European Environment Agency (EEA)
}
\keyword{datasets}

