% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DasyMapR-GenUt.R
\docType{methods}
\name{dasymapPlot,EtrsSurface,numeric-method}
\alias{dasymapPlot,EtrsSurface,numeric-method}
\title{a utility for plotting EtrsSurface combining attributes}
\usage{
\S4method{dasymapPlot}{EtrsSurface,numeric}(aEtrsSurface, col.value)

\S4method{dasymapPlot}{EtrsSurface,numeric}(aEtrsSurface, col.value)
}
\arguments{
\item{aEtrsSurface}{EtrsSurface.}

\item{col.value}{numeric.}

\item{sppdf}{EtrsSurface}

\item{column}{of the class value}
}
\value{
a plot

a leaflet plot
}
\description{
produce a 5 class choroplethe map for en ETRS surface

(lealet version)
}

