% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DasyMapR-package.R
\docType{package}
\name{DasyMapR-package}
\alias{DasyMapR}
\alias{DasyMapR-package}
\title{DasyMapR}
\description{
The process of dasymetric mapping is the transformation of data
from a set of arbitrary source zones to a dasymetric map via the overlay of
the source zones with an ancillary data set.This package contains a number of
fucntions that assist preparing vector and raster source and ancillary datasets
for intelligent dasymetric mapping, performs the dasymetric calculations, and
then generates a floating point output raster of revised density.
TO BE NOTICED : It uses the ETRS89-LAEA geographic grid to do the job.
}

