% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ETRS.R
\docType{class}
\name{ETRS-class}
\alias{ETRS}
\alias{ETRS-class}
\title{Virtual Parent Class for ETRS objects}
\value{
One cannot create an ETRS instance inasmuch as ETRS is VIRTUAL class
}
\description{
The ETRS class contains (extends) the Spatial Class. As a result of that has the slots of the parent class : @bbox and proj4string.
}
\section{Slots}{

\describe{
\item{\code{etrs.cell.codes.columns}}{character: a character vector c("CELLCODE","EASTOFORIGIN","NORTHORIGIN") The Enviromental European Agency suggests to be provided, except "CELLCODE", for computations and analysis the lower left corner coordinate in meters, formatted as two separate integers attribute values named "EOFORIGIN" and "NOFORIGIN"}

\item{\code{VIRTUAL.}}{Actually only to declare the ETRS class abstract}
}}

