% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-EtrsSurface.R
\name{etrsCellCenter}
\alias{etrsCellCenter}
\title{The polygon in which the center of the cell yields the attribute to assign to the cell.}
\usage{
etrsCellCenter(the.etrs.grid = "EtrsGrid")
}
\arguments{
\item{the.etrs.grid}{an EtrsGrid}
}
\value{
Returns centroids of an etrsgrid
}
\description{
The polygon in which the center of the cell yields the attribute to assign to the cell.
}

