% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ETRS.R
\docType{methods}
\name{EtrsCheckCodeColumns,ETRS-method}
\alias{EtrsCheckCodeColumns,ETRS-method}
\title{checks if colum names data are etrs table.codes}
\usage{
\S4method{EtrsCheckCodeColumns}{ETRS}(object)
}
\arguments{
\item{Etrs}{an etrs object}
}
\description{
Is the slot etrs.table.cell.code identical to data colum names?
}
\examples{

}

