% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-DasymetricSurface.R
\docType{methods}
\name{etrsDasymetric2Ancillary,EtrsDasymetricSurface-method}
\alias{etrsDasymetric2Ancillary,EtrsDasymetricSurface-method}
\title{A method to convert a DasyMetric surface to Ancillary}
\usage{
\S4method{etrsDasymetric2Ancillary}{EtrsDasymetricSurface}(dasymetric.surface)
}
\arguments{
\item{dasymetric.surface}{EtrsDasymetricSurface.}
}
\value{
EtrsAncillarySurface
}
\description{
A method to convert a DasyMetric surface to Ancillary
}
\examples{

}

