% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-DasymetricSurface.R
\docType{methods}
\name{etrsDasymetric2Raster,EtrsDasymetricSurface-method}
\alias{etrsDasymetric2Raster,EtrsDasymetricSurface-method}
\title{A simple method to produce a floating point raster}
\usage{
\S4method{etrsDasymetric2Raster}{EtrsDasymetricSurface}(dasymetric.surface)
}
\arguments{
\item{dasymetric.surface}{EtrsDasymetricSurface.}

\item{dasymetric.surface}{EtrsDasymetricSurface.}
}
\value{
a Raster object and a geo tif file in working directory

a Raster object and a geo tif file in working directory
}
\examples{

data(DASY_GDP.rda)
par(mar = c(0.1, 0.1, 0.1, 0.1))
DASY_GPD_RASTER <- etrsDasymetric2Raster(dasymetric.surface = DASY_GPD)
rw.colors <- grey.colors
image(DASY_GPD_RASTER, col = rw.colors(5))
A simple method to produce a rfloating point aster floating point

}

