% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DasyMapR-GenUt.R
\docType{methods}
\name{etrsDasymetric2Source,EtrsDasymericSurface-method}
\alias{etrsDasymetric2Source,EtrsDasymericSurface-method}
\title{simple utillity converts EtrsDasymetric 2  EtrsSourceSurface}
\usage{
\S4method{etrsDasymetric2Source}{EtrsDasymericSurface}(d)
}
\arguments{
\item{d}{EtrsDasymericSurface.}
}
\value{
EtrsSourceSurface
}
\description{
simple utillity converts EtrsDasymetric 2  EtrsSourceSurface
}

