% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-DasymetricSurface.R
\docType{methods}
\name{EtrsDasymetricSurface,EtrsSourceSurface,EtrsAncillarySurface-method}
\alias{EtrsDasymetricSurface,EtrsSourceSurface,EtrsAncillarySurface-method}
\title{EtrsDasymetricSurface method}
\usage{

  \S4method{EtrsDasymetricSurface}{EtrsSourceSurface,EtrsAncillarySurface}(input.surface.grided,
  ancillary.grided, ...)
}
\arguments{
\item{input.surface.grided}{EtrsSourceSurface.}

\item{ancillary.grided}{EtrsAncillarySurface.}

\item{actuall.value}{logical.}
}
\description{
EtrsDasymetricSurface method
}

