% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-EtrsGrid.R
\docType{methods}
\name{etrsGrid,SpatialPolygonsDataFrame,numeric-method}
\alias{etrsGrid,SpatialPolygonsDataFrame,numeric-method}
\title{the etrsGrid default method}
\usage{

  \S4method{etrsGrid}{SpatialPolygonsDataFrame,numeric}(obj = "SpatialPolygonsDataFrame",
  cell.size = "numeric")
}
\arguments{
\item{obj}{SpatialPolygonsDataFrame.}

\item{cell.size}{numeric. @param cell.size a numeric one of the resolutions of the grid  that is 0.5m, 1m, 2.5m, 5m, 10m, 25m, 50m, 100m, 250m, 500m, 1Km, 2.5Km, 5Km, 10Km, 25Km, 50Km, 100Km.}
}
\value{
EtrsGrid class object
}
\description{
the etrsGrid default method
}
\examples{
#' x <- round(runif(1, min = 4320200, max = 4322000), 1)
y <- round(runif(1, min = 3210000, max = 3212000), 1)
xy <- cbind(x, y)
size = round(runif(1,min = 1500,max = 2000),0)
pl1 <-
  Polygons(list(Polygon(cbind(
    c(x,x + size,x + size,x,x),c(y,y,y + size,y + size,y)
  ))),round(runif(1,min = 1,max = 100),0))
x <- x + size
pl2 <-
  Polygons(list(Polygon(cbind(
    c(x,x + size,x + size,x,x),c(y,y,y + 2 * size,y + 2 * size,y)
  ))),round(runif(1,min = 1,max = 100),0))
x<-x-size
y<-y+size
pl3 <-
  Polygons(list(Polygon(cbind(
    c(x,x + size,x + size,x,x),c(y,y,y +  size,y +  size,y)
  ))),round(runif(1,min = 1,max = 100),0))
sps <-
  SpatialPolygons(list(pl1,pl2,pl3),proj4string = CRS("+init=epsg:3035"))
df <-
  data.frame(val=c("R5","R40","R80"),row.names = sapply(slot(sps, "polygons"), function(x)
    slot(x, "ID")))
sps<-SpatialPolygonsDataFrame(sps,data = df)
the.etrs.grid<-etrsGrid(sps,1000)
slotNames(the.etrs.grid)
the.etrs.grid@the.grid.name
head(the.etrs.grid@data,3)
plot(the.etrs.grid)
title(paste("the etrs grid",the.etrs.grid@the.grid.name))
}

