% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-EtrsGrid.R
\name{etrsGrid2Spdf}
\alias{etrsGrid2Spdf}
\title{Converts an etrsgrid obj to standard SpatiaPolygonsDataFrame}
\usage{
etrsGrid2Spdf(the.etrs.grid = "etrsGrid")
}
\arguments{
\item{the.etrs.grid}{}
}
\value{
SpatialPolygonsDataFrame object
}
\description{
It is usefull for rGeos that doesn't support coercion. ALso to be used for coercion
}
\examples{


}

