% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-EtrsPoint.R
\docType{class}
\name{EtrsPoints-class}
\alias{EtrsPoints}
\alias{EtrsPoints-class}
\title{EtrsPoints class extends \code{"ETRS"} parrent class}
\description{
Represents a set of lower corner grid points of the geographical grid ETRS89-LAEA
}
\section{Slots}{

\describe{
\item{\code{SpatialPointsDataFrame}}{: an object of SPatialPointsDataFrame}

\item{\code{etrs.cell.codes.columns}}{: a character object that keeps the colnames of the dataframe}
}}
\examples{
# Test for EtrsPoint
bb <- bbox(cbind(c(4321000, 4323000), c(3210000,3212000)))
cell.size = 1000
# Calculate the limits of the grid
bb.outer.limits <-
cbind((floor(bb[, 1] / cell.size)) * cell.size, (ceiling(bb[, 2] / cell.size))*cell.size)
eastings <- seq(from = bb.outer.limits[1, 1], to = bb.outer.limits[1, 2], by = cell.size)
#Calculates northings
northings <- seq(from = bb.outer.limits[2, 1], to = bb.outer.limits[2, 2], by = cell.size)
eastings.northings <- as.matrix(expand.grid(eastings, northings))
low.corner.cell.etrs.points<-etrsPoints(eastings.northings = eastings.northings,cell.size=10000)
plot(low.corner.cell.etrs.points)
text(eastings.northings[,1]+200,eastings.northings[,2],low.corner.cell.etrs.points[["NORTHOFORIGIN"]])
text(eastings.northings[,1],eastings.northings[,2]+200,low.corner.cell.etrs.points[["EASTOFORIGIN"]],srt=90)
title("Lower Cell Corner Etrs Points using ETRS-LAEA")
}

