% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-EtrsPoint.R
\docType{methods}
\name{etrsPoints,matrix,numeric-method}
\alias{etrsPoints,matrix,numeric-method}
\title{\code{"etrsPoints"} is the \code{"EtrsPoints"} object constructor}
\usage{
\S4method{etrsPoints}{matrix,numeric}(eastings.northings = "matrix",
  cell.size = "numeric")
}
\arguments{
\item{eastings.northings}{: a matrix of coordinates}

\item{cell.size}{: a numeric object represnts the cell size of the grid}
}
\value{
An EtrsPoints object
}
\description{
\code{"etrsPoints"} is the \code{"EtrsPoints"} object constructor
}
\examples{
# Test for EtrsPoint
bb <- bbox(cbind(c(4321000, 4323000), c(3210000,3212000)))
cell.size = 1000
# Calculate the limits of the grid
bb.outer.limits <- cbind((floor(bb[, 1]/cell.size))*cell.size,(ceiling(bb[,2]/cell.size))*cell.size)
eastings <- seq(from = bb.outer.limits[1, 1], to = bb.outer.limits[1, 2], by = cell.size)
#Calculates northings
northings <- seq(from = bb.outer.limits[2, 1], to = bb.outer.limits[2, 2], by = cell.size)
eastings.northings <- as.matrix(expand.grid(eastings, northings))
low.corner.cell.etrs.points<-etrsPoints(eastings.northings = eastings.northings,cell.size=1000)
plot(low.corner.cell.etrs.points)
text(eastings.northings[,1]+200,eastings.northings[,2],low.corner.cell.etrs.points[["NORTHOFORIGIN"]])
text(eastings.northings[,1],eastings.northings[,2]+200,low.corner.cell.etrs.points[["EASTOFORIGIN"]],srt=90)
title("Lower Cell Corner Etrs Points using ETRS-LAEA")
}

