% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-DasymetricSurface.R
\name{etrsPropWeightedValue}
\alias{etrsPropWeightedValue}
\title{Proportional and weighted calculation}
\usage{
etrsPropWeightedValue(input.surface.grided = "EtrsSourceSurface",
  ancillary.grided = "EtrsAncillarySurface")
}
\arguments{
\item{ancillary.grided}{an EtrsAncillarySurface}

\item{etrs.grided.source.sur}{EtrsSourceSurface}
}
\value{
a EtrsDasymetricSurface
}
\description{
the cell takes also a proportionally
calculated value, but this value is weighted for each cell, according to an external
variable (e.g. population). This method can be applied to improve the territorial
distribution of a socioeconomic indicator.
Cell value = W c Σ ( V i * Share i )
V i = Value of unit i
Share i = Share of unit i within the cell,
 W c = weight assigned to cell
}
\examples{

}

