% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ETRS.R
\docType{methods}
\name{etrsReverseCellCode,data.frame,numeric-method}
\alias{etrsReverseCellCode,data.frame,numeric-method}
\title{Reversing etrs cell codew from resoloution easting northings to res northing eastings}
\usage{
\S4method{etrsReverseCellCode}{data.frame,numeric}(df = "data.frame",
  cell.code.col = "numeric")
}
\arguments{
\item{df}{.  A data frame that has a Etrs reversed ceel code column}

\item{cell.code.col}{. Define the column of the reversed cellcode}
}
\value{
a data frame with CELLCODE and ID = CELLCODE
 @description It 's possible to find out, out there data not comabatible to Inspire.
 so reversin is a option. Such an example the GEOASTAT_grid_EU_POP .... files
the method also set the data frame row names as etrs cell codes. As a result of it it can be used to as a spatial etrs grid
}
\description{
Reversing etrs cell codew from resoloution easting northings to res northing eastings
}
\examples{

}

