% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-EtrsSurface.R
\docType{methods}
\name{etrsSurface,SpatialPolygonsDataFrame,missing,missing,numeric-method}
\alias{etrsSurface,SpatialPolygonsDataFrame,missing,missing,numeric-method}
\title{Etrs (source or Ancillary) surface creation}
\usage{

  \S4method{etrsSurface}{SpatialPolygonsDataFrame,missing,missing,numeric}(input.surface,
  over.method.type, surface.value.col, cell.size, ...)
}
\arguments{
\item{over.method.type}{"PropCal" in proportional calculation the cell takes a calculated value depending on the values of the units falling inside and their share within the cell. This method seems very appropriate for countable variables}

\item{surface.value.col}{the number of colum that keeps  the relative density of a cell with land-cover some type}

\item{cell.size:}{one of the resolutions of the grid  that is 0.5m, 1m, 2.5m, 5m, 10m, 25m, 50m, 100m, 250m, 500m, 1Km, 2.5Km, 5Km, 10Km, 25Km, 50Km, 100Km}
}
\description{
Depending on each type of indicator or variable to be integrated within the reference grid, a different type of integration should be decided and tested. Besides the method finally chosen to integrate, it is important to highlight that indicator figures given by area unit, e.g. by square kilometre, should be converted considering that each cell has a total area of 1 km 2.
}
\seealso{
etrsSurface.default, etrsMaxArea, etrsPropCal
}

