% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-EtrsSurface.R
\name{etrsSurface2Spdf}
\alias{etrsSurface2Spdf}
\title{Simple tool to convert a etrsSurface Object to Spatial}
\usage{
etrsSurface2Spdf(the.etrs.surface = "EtrsSurface")
}
\arguments{
\item{the.etrs.surface}{}
}
\value{
SpatialPolygonsDataFrame
}
\description{
Simple tool to convert a etrsSurface Object to Spatial
}

