% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-EtrsSurface.R
\docType{methods}
\name{etrsSurfacePar,SpatialPolygonsDataFrame,ANY,numeric,numeric-method}
\alias{etrsSurfacePar,SpatialPolygonsDataFrame,ANY,numeric,numeric-method}
\title{etrsSurfacePar}
\usage{

  \S4method{etrsSurfacePar}{SpatialPolygonsDataFrame,ANY,numeric,numeric}(input.surface,
  over.method.type, surface.value.col, cell.size)
}
\arguments{
\item{input.surface}{SpatialPolygonsDataFrame.}

\item{over.method.type}{ANY.}

\item{surface.value.col}{numeric.}

\item{cell.size}{numeric.}
}
\value{
an EtrsSurface
}
\description{
etrsSurfacePar
}

