% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ETRS.R
\docType{methods}
\name{EtrsTableCodes,matrix,numeric-method}
\alias{EtrsTableCodes,matrix,numeric-method}
\title{A method that produces a data frame of cell codes and lower corner easting
northings}
\usage{
\S4method{EtrsTableCodes}{matrix,numeric}(eastings.northings, cell.size)
}
\arguments{
\item{eastings.northings}{a matrix populated with lower corner grid coordinates}

\item{cell.size:}{one of the resolutions of the grid  that is 0.5m, 1m, 2.5m, 5m, 10m, 25m, 50m, 100m, 250m, 500m, 1Km, 2.5Km, 5Km, 10Km, 25Km, 50Km, 100Km}
}
\value{
dataframe
}
\description{
A method that produces a data frame of cell codes and lower corner easting
northings
}
\examples{
bb <- bbox(cbind(c(4321000, 4323000), c(3210000,3212000)))
cell.size = 1000
# Calculate the limits of the grid
bb.outer.limits <-
 cbind((floor(bb[, 1] / cell.size)) * cell.size, (ceiling(bb[, 2] / cell.size)) * cell.size)
# Calculates eastings
eastings <-
 seq(from = bb.outer.limits[1, 1], to = bb.outer.limits[1, 2], by = cell.size)
# Calculates northings
northings <-
 seq(from = bb.outer.limits[2, 1], to = bb.outer.limits[2, 2], by = cell.size)
eastings.northings <- as.matrix(expand.grid(eastings, northings))
low.corner.cell.etrs.points<-etrsPoints(eastings.northings = eastings.northings,cell.size =1000)
df<-EtrsTableCodes(eastings.northings,1000)
df
}

