% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-EtrsSurface.R
\name{EtrsTransform}
\alias{EtrsTransform}
\title{Transforms from the current CRS to ETRS-LAEA}
\usage{
EtrsTransform(obj)
}
\arguments{
\item{obj}{}
}
\value{
an EtrsSurface in ETRS CRS
}
\description{
Transforms from the current CRS to ETRS-LAEA
}

