% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-EtrsSurface.R
\name{joinMaxAreaSurfaceDataFrames}
\alias{joinMaxAreaSurfaceDataFrames}
\title{Simple tool to join etrsSurface with DataFrame}
\usage{
joinMaxAreaSurfaceDataFrames(the.surface = "SpatialPolygonsDataFrame",
  the.EtrsSurface = "EtrsSurface")
}
\arguments{
\item{the.surface}{An input Surface}

\item{the.EtrsSurface}{an EtrsSurface object}
}
\value{
an EtrsSurface
}
\description{
Simple tool to join etrsSurface with DataFrame
}

