% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DasyMapR-data-documentation.R
\docType{data}
\name{nama_10r_3gdp}
\alias{nama_10r_3gdp}
\title{Gross domestic product (GDP) at current market prices by NUTS 3 regions}
\format{An object of class \code{data.frame} with 10980 rows and 4 columns.}
\usage{
nama_10r_3gdp
}
\description{
Downloaded from Eurostat via eurostat package. Just a sample  elected from the original dataset for NUTS65 and year 2001
}
\note{
© European Union, 1995-2016 Reuse is authorised, provided the source is acknowledged. The reuse policy of the European Commission is implemented by a Decision of 12 December 2011
}
\author{
Eurostat
}
\keyword{datasets}

