% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DasyMapR-data-documentation.R
\docType{data}
\name{NUTS3_OCMG}
\alias{NUTS3_OCMG}
\title{NUTS3_OMG : Boundaries of Greece administartive Units of level NUTS3}
\format{A Spatial polygons data frame}
\source{
\url{http://geodata.gov.gr/en/dataset/6deb6a12-1a54-41b4-b53b-6b36068b8348/resource/3e571f7f-42a4-4b49-8db0-311695d72fa3/download/nomoiokxe.zip}
}
\usage{
NUTS3_OCMG
}
\description{
the data set was provided by Hellinic Organiztion of Cadastral and Mapping as the official
boundaries for   perfectures (NUTS3) greece. The data is under the creative commons licences 3.0 and is .shp file.  The db file also contains some population data.
}
\note{
Share — copy and redistribute the material in any medium or format Adapt — remix, transform, and build upon the material for any purpose, even commercially. The licensor cannot revoke these freedoms as long as you follow the license terms. CC 3.0
}
\keyword{datasets}

