% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DasyMapR-data-documentation.R
\docType{data}
\name{NUTS_2013_01M_EL}
\alias{NUTS_2013_01M_EL}
\title{Administrative or Statistical unit NUTS2013}
\format{An object of class \code{SpatialPolygonsDataFrame} with 70 rows and 4 columns.}
\usage{
NUTS_2013_01M_EL
}
\description{
At the beginning of the 1970s, Eurostat set up the NUTS classification as a single, coherent system for dividing up the EUs territory in order to produce regional statistics for the Community. For around thirty years, implementation and updating of the NUTS classification was managed under a series of "gentlemen s agreements" between the Member States and Eurostat.However, sometimes national interests require changing the regional breakdown of a country. When this happens the country concerned informs the European Commission about the changes.

 @format Spatial Polygons Data Frames
 @source \url{http://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units/nuts#nuts13}
 @author Eurostat
 @notes Copyright notice EN: © EuroGeographics for the administrative boundaries
}
\keyword{datasets}

