% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DasyMapR-data-documentation.R
\docType{data}
\name{NUTSV9_LEAC}
\alias{NUTSV9_LEAC}
\title{NUTSV9_LEAC: Administrative land accounting units in Lambert Equal Area}
\format{A Spatial polygons data frame}
\source{
\url{http://epp.eurostat.ec.europa.eu/portal/page?_pageid=2254,62148876,2254_62153824&amp;_dad=portal&amp;_schema=PORTAL}
}
\usage{
NUTSV9_LEAC
}
\description{
GISCO administrative boundaries (NUTS) v9 generalised using the 1 km reference grid for the Land cover accounts project (LEAC) Administrative land accounting units are used to allocate land cover changes and to relate socio-economic processes to land cover dynamics. The GISCO database (Eurostat) provides a medium-scale layer for regional administrative boundaries (NUTS) covering the entire EU territory. The hierarchy of administrative land accounting units allows the analysis of the data at various scales, from NUTS 3 (province level) to NUTS 0 (country level). A mixed level, combining NUTS 3 and NUTS 2 was defined in order to homogenize the size of the units.
}
\note{
Rights: EEA standard re-use policy, unless otherwise indicated, re-use of content on the EEA website for commercial or non-commercial purposes is permitted free of charge, provided that the source is acknowledged (http://www.eea.europa.eu/legal/copyright). Copyright holder: European Environment Agency (EEA).
}
\author{
EEA
}
\keyword{datasets}

