% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ETRS.R
\docType{methods}
\name{EtrsCellCodes,matrix,numeric,missing-method}
\alias{EtrsCellCodes,matrix,numeric,logical-method}
\alias{EtrsCellCodes,matrix,numeric,missing-method}
\title{A cell code identifier for ETRS}
\usage{
\S4method{EtrsCellCodes}{matrix,numeric,missing}(eastings.northings, cell.size)

\S4method{EtrsCellCodes}{matrix,numeric,logical}(eastings.northings, cell.size,
  NE)
}
\arguments{
\item{eastings.northings}{matrix.}

\item{cell.size}{numeric.}

\item{eastings.northings}{matrix.}

\item{cell.size}{numeric.}
}
\value{
etrs cell code vector

etrs cell code vector
}
\description{
Define the cell size prefix

Define the cell size prefix
}
\examples{
bb <- bbox(cbind(c(4321000, 4323000), c(3210000,3212000)))
cell.size = 1000
# Calculate the limits of the grid
bb.outer.limits <-  cbind((floor(bb[, 1] / cell.size)) * cell.size, (ceiling(bb[, 2] / cell.size)) * cell.size)
# Calculates eastings
eastings <-seq(from = bb.outer.limits[1, 1], to = bb.outer.limits[1, 2], by = cell.size)
# Calculates northings
northings <-seq(from = bb.outer.limits[2, 1], to = bb.outer.limits[2, 2], by = cell.size)
eastings.northings <- as.matrix(expand.grid(eastings, northings))
low.corner.cell.etrs.points<-etrsPoints(eastings.northings = eastings.northings,cell.size =1000)
EtrsCellCodes<-EtrsCellCodes(eastings.northings,cell.size=1000)
 
bb <- bbox(cbind(c(4321000, 4323000), c(3210000,3212000)))
cell.size = 1000
# Calculate the limits of the grid
bb.outer.limits <-  cbind((floor(bb[, 1] / cell.size)) * cell.size, (ceiling(bb[, 2] / cell.size)) * cell.size)
# Calculates eastings
eastings <-seq(from = bb.outer.limits[1, 1], to = bb.outer.limits[1, 2], by = cell.size)
# Calculates northings
northings <-seq(from = bb.outer.limits[2, 1], to = bb.outer.limits[2, 2], by = cell.size)
eastings.northings <- as.matrix(expand.grid(eastings, northings))
low.corner.cell.etrs.points<-etrsPoints(eastings.northings = eastings.northings,cell.size =1000)
EtrsCellCodes<-EtrsCellCodes(eastings.northings,cell.size=1000,NE=TRUE)
}

