% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DasyMapR-GenUt.R
\docType{methods}
\name{raster2Ancillary,raster,numeric,numeric-method}
\alias{raster2Ancillary,raster,numeric,numeric-method}
\title{create a EtrsAncillary Surface from a Raster oBject}
\usage{
\S4method{raster2Ancillary}{raster,numeric,numeric}(aRaster, cell.size,
  attr_divisor = 1)
}
\arguments{
\item{aRaster}{raster.}

\item{cell.size}{numeric}

\item{attr_divisor}{numeric.}
}
\value{
EtrsAncillarySurface
}
\description{
create a EtrsAncillary Surface from a Raster oBject
}

