% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DasyMapR-GenUt.R
\docType{methods}
\name{wgsTransform,Spatial-method}
\alias{wgsTransform,Spatial-method}
\title{Transforms to wgs crs}
\usage{
\S4method{wgsTransform}{Spatial}(obj)
}
\arguments{
\item{EtrsSurface}{}
}
\value{
EtrsSurface
}
\description{
usefull for Google,Osm e.t.c.
}

